feat: unit test for error page by tdd method

+++ src/main/java/com/glenncai/studentsystem/controller/GradebookController.java
@@ -47,6 +47,10 @@ public class GradebookController {
 
   @GetMapping("/delete/student/{id}")
   public String deleteStudent(@PathVariable int id, Model m) {
+    if (!studentAndGradeService.checkIfStudentExists(id)) {
+      return "error";
+    }
+
     studentAndGradeService.deleteStudent(id);
     Iterable<CollegeStudent> collegeStudents = studentAndGradeService.getGradebook();
     m.addAttribute("students", collegeStudents);
